pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "bark"
version = "0.1.0"
description = "BARK RL/IL training pipeline for dog leg dependency learning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "gymnasium>=0.29",
    "gymnasium[mujoco]>=0.29",
    "mujoco>=3.0",
    "stable-baselines3>=2.2",
    "wandb>=0.16",
    "imitation>=0.5",
    "scipy>=1.11",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[tool.setuptools.packages.find]
where = ["."]
include = ["envs*", "train*", "data*"]
namespaces = false
